+++ NEWS.md
@@ -1,3 +1,14 @@
-# BCyto 1.0.1
+# BCyto 1.0.2 (02/02/2022)
+
+* AutoSpill fixed for Windows.
+* File paths fixed for Windows.
+* Compensation matrices application fixed.
+* Scale tool fixed and improved.
+* Export button on Proliferation tab fixed.
+* Re-generation of tSNE plots fixed.
+* Download test button removed.
+
+
+# BCyto 1.0.1 (21/01/2022)
 
 * Download test data button fixed for Mac OS.